regex for station

+++ gradski-prevoz/Location.h
@@ -13,8 +13,8 @@ public:
 	~Location();
 	inline void setLatitude(double l) { latitude = l; }
 	inline void setLongitude(double l) { longitude = l; }
-	inline int getLatitude() const { return latitude; }
-	inline int getLongitude() const { return longitude; }
+	inline double getLatitude() const { return latitude; }
+	inline double getLongitude() const { return longitude; }
 	friend ostream& operator<<(ostream &it, const Location &l);
 
 private:
